Allow dbstop to set breakpoints outside of the function currently being debugged.
* debug.cc (parse_dbfunction_params): If first input is a string, use atoi to
decide whether it is a real string (function name) or a line number for a
breakpoint passed in as a string.  If it is a line number, grab the function
name from the function currently being debugged.  Issue an error message about
struct input not being implemented if a struct is found.  Skip struct inputs
when they are encountered in the list of breakpoint line numbers.

* debug.cc (Fdbstatus): Add note about breakpoint list being trimmed when in
debug mode inside a function.