nextpow2: compute for each element of input instead of its length (bug #44291)
* general/nextpow2.m: when input is not a scalar, nextpow2() was using
length(x).  Not only is this Matlab incompatible, it is also not very
useful in the light of nd signals where input will be the length of each
dimension.  It would also break very bad when the user wants to pass a
vector that has length 1 in which case it would use the first element of
the vector. This commit changes it so that it computes nextpow2 for each
element of the input.  When the previous behaviour is intended,
`nextpow2 (length (x))' should be used.
* NEWS: make note of this backwards incompatible change.