No terminal window when starting Octave from desktop menu (bug #44349)
* etc/icons/octave.desktop.in: the Terminal key on a desktop file, controls
whether the program runs on a terminal window.  While this is true for Octave
on its own, we can guess that if the user is using desktop menus he wants the
Octave GUI.  Anyway, as it was, it was simply opening a terminal window and
run Octave from there which was useless.  Users wanting a icon for a terminal
window with Octave (no-gui), will have to create it themselves.