Script file or eval code are not interactive session so keep Octave quiet.
* libinterp/octave.cc: add a new static variable an_octave_program to keep
track when this is called for an Octave program and not for an Octave
interactive session. The main point of this patch is so that
octave_startup_message() on octave_execute_interpreter() is not called.
A continuation work from csets 6ba5f1ff041e and f3ac54ac2c6a which also
aimed at doing this.  The idea is to have an_octave_program inherit other
features of a non-interactive session (such as being canceled by forcing
an interactive session or not starting readline).