Convert octave integer scalars into closest java type (bug #44882)
* libinterp/octave-value/ov-java.cc (unbox): when calling a java method,
octave core types should be converted automatically into equivalent java
types.  In the case of integers, we must pick the closest one but at the
moment all were being converted to Integer (int32_t).  This fixes it, even
if it meas to lose data (as is the case in some of the added tests).  Also
reduce some code duplication by using macros since the code is similar to
the code for float and doubles.