hist: reduce memory usage by avoiding conversion to double (bug #47707)
* hist.m: do not convert whole input array to double.  Instead, convert only
the pieces that are needed (max and min values), and floor the cutoff point
so they are placed in the right place when sorted (for N > 30 only).  There
was no issue with overflow as the comments suggested (at least not anymore).
Further reduce memory by avoiding intermediate variables.  Add more logic so
that NBINS or specified X can be of class single.  Add tests for all this.