inputParser: new method addParameter --- just like addParamValue (bug #45367)
* inputParser.m: for unknown reasons, addParamValue is no longer recommended
in Matlab and instead addParameter should be used.  The syntax and behaviour
seems to be exactly the same, the only difference being that addParameter
supports one more option at the end.  We do not add support for that extra
option yet though.  Also, we do not deprecate or recommend against the use
of addParamValue because it works fine and we care more about backwards
compatibility.