Overhaul graphics implementation of hold() (bug #43559).
* NEWS: Announce possible color changes in plots relying on "hold on".

* hold.m: Deprecate "hold all" in docstring.  Eliminate use of __hold_all__
graphics property.  Update %!demo to use "hold on" only.

* ishold.m: Put input validation first.  Use switch statement rather than
if/else tree for clarity.

* graphics.in.h: Remove hidden __hold_all__ graphics property of axes.

* graphics.cc (axes::properties::set_defaults): Remove use of __hold_all__.

* newplot.m: Delete kluge use of __next_line_color__ and __next_line_style__.
Set colororderindex and linestyleorderindex to 1 on new plots.

* __next_line_color__.m: Completely rewrite to use the axes graphic property
"colororderindex" to keep track of which color should be used next.  Also
increment "linestyleorderindex" when number of colors used rolls over so that
colors will repeat with the next linestyle in the series.  Update BIST test.

* __next_line_style__.m: Completely rewrite to use the axes graphic property
"linestyleorderindex" to keep track of which linestyle should be used next.

* imagesc.m, axis.m, legend.m: Replace "hold all" with "hold on" in all demos.