xzip.cc: replace throwing exceptions in destructors (bug #48640)
* xzip.cc: we can't throw exceptions in destructors.  Replace them with
a close method that can throw while destructor closes without throwing.
If we want to close something cleanly, then we do want to know of any
failure and catch failures.  If we are not closing it cleanly, then we
either don't care or are already handling another expcetion, and either
way we don't want to handle an exception.