Don't error out if RANGE specification for dlmread contains large values.
* dlmread.cc (idx_max): Redefine static const variable that holds the maximum
row or column idx for a range specification to be max octave_idx_type - 1.
* dlmread.cc (idx_max_dbl): New static const variable that holds the maximum
possible row or column idx for a range specification as a double value.
* dlmread.cc (parse_range_spec): Use new idx_max constant.
Use std::min (x, idx_max_dbl) to eliminate overflow when casting double to
octave_idx_type.
* dlmread.cc (Fdlmread): Use new idx_max constant.  After parsing range
specification, return an empty matrix immediately if range is empty.  When
skipping lines at beginning of file, look at both the number of lines to skip
and the output of getline; The EOF may be reached before the number of lines to
skip.