don't use singleton for symbol_table
This is the first of a series of changes to make the symbol table a
part of the interpreter instead of a global object.  These changes
also aim to simplify the implementation of symbol table so that it is
easier to understand and modify.

  * Functions now own their scope (workspace) data.

  * The list of subfunctions is contained in the scope rather than a
    global list.

* symtab.h, symtab.cc (class symbol_table): Don't use singleton
pattern.

* interpreter.h, interpreter.cc (interpreter::m_symbol_table): New
data member.
(interpreter::~interpreter): Don't set instance to 0.

* interpreter-private.h, interpreter-private.cc
(__get_symbol_table__): New function.

Change all uses of call_stack to access call_stack object from the
interpreter.