make symbol table context info mostly local to symbol table scopes
The context of a symbol table is used to handle recursion.  A new
context is created in a scope when a function is called recursively.
In the past, we managed the context globally.  Now it is local to the
scope and does not need to be passed to functions that access symbol
record data.