delete filemarker function and corresponding internal variable
This variable stored the character used to separate function,
subfunction, and nested function names.  I see no reason to keep it.
The name 'filemarker' does not clearly indicate what this variable is
about.  Changing the value from the default '>' would
likely create a great deal of confusion, so I doubt that it was ever
set to anything different.  We don't have a similar way to change the
'@' or '+' characters used to prefix class or package names.  So
rather than deprecate, I'm simply removing it.

* input.h, input.cc (Vfilemarker, Ffilemarker): Delete internal
variable and function.  Change all uses of Vfilemarker to be '>'
instead.