solve installation info initialization problem differently
Back out changesets 893344cee100 and 69fc8935020b.

* defaults.h, defaults.cc: Instead of storing installation info like
include_dir, data_dir, info_dir, etc., in file-scope static variables
or a class object owned by the interpreter, store the initialized
values as constants inside the functions that return them.  This way
they are initialized on demand and we don't have to worry about
initialization order of static data.  It's OK for these to be static
because even though they may require some computation to intialize
(looking at environment variables or substituting the value of
OCTAVE_HOME), they are constants for any given installation of
Octave and will be the same for any instantiation of the interpreter.

Only include defaults.h in files that actually need it.