Improve fread speed by a further 40% when SKIP parameter used (bug #54100).
* oct-stream.cc (stream::fread): Check stream_ok immediately and return from
function if it is not.  Reformat comments to 80 character width.  Declare
new variable cur_pos to hold current position within file.  Remove call to
tellg, and use arithmetic based on cur_pos and number of iems read to determine
position in file.  Eliminate unneccesary test on stream is after seekg as it
will be done as part of the while loop.