always provide definitions for deprecated functions
Previously, deprecated functions that were defined in source files
instead of as inline functions in header files were not available even
when OCTAVE_USE_DEPRECATED_FUNCTIONS was defined (declarations were
made available, but the definitions were not present, so using them
was impossible).

* input.cc, utils.cc, oct-base64.cc: Always provide
definitions of deprecated functions that are defined in source files.

* lo-specfun.cc: Delete wrappers for functions no longer exist.

* input.h, input.cc (set_default_prompts): Delete wrapper for functon
that no longer exists.

* MArray-i.cc, MArray-s.cc: Delete instantiations of functions that no
longer exist.