Correctly handle corner cases in dlmread. Fix handling of <number>i|j<extra text> Fix handling of <Inf|NA|NaN><extra text>
* dlmread.cc: For purely complex numbers, check that there is no trailing
character after i|j.  If there is, the field failed to parse correctly.
Advance the column counter with j++ and leave data initialized to empty_value.
For non-finite values, check that there is no extra text beyond the parsed
text.  If there is, the field failed to parse correctly.  Advance the column
counter with j++ and leave data initialized to empty_value.
If octave_read_double() parsing fails use j++ to advance the column counter and
leave data initialized to empty_value.  Add BIST tests for purely complex
numbers with trailing characters and for non-finite values with trailing
characters.