doc: Add version number to top title of HTML manual for clarity.
Currently only "GNU Octave" is the top title of each HTML manual, cf.

- https://octave.org/doc/v4.0.0/
- https://octave.org/doc/v4.2.0/
- https://octave.org/doc/v4.4.0/
- https://octave.org/doc/interpreter/

and the version can be seen later in the Texinfo #top node:

- https://octave.org/doc/v4.0.0/index.html#Top
...
- https://octave.org/doc/interpreter/index.html#Top

To make it easier to identify the manual version not only from the URL, what in
the case of the latest manual "interpreter" is impossible, put it inside the
top title.

This change does not affect the PDF manual.