figure.m: Accept structures of property/value pairs (bug #44672).
* figure.m: Change input validation to look for a numeric scalar argument as
the first argument.  If it is then validate that it is either a figure handle
or an integer for a new figure.  Use cellfun to determine which inputs in
varargin are structs and remove them before calculating whether the number
of property/value pairs is a multiple of 2.  Adjust BIST tests to pass.