Fix failing BIST tests and oddities when gnuplot is not present (bug #55023).
* acinclude.m4 (OCTAVE_PROG_GNUPLOT): When gnuplot is not found, set GNUPLOT to
the default value of "gnuplot" rather than an empty string.  Also set variable
GNUPLOT_BINARY which is only used for reporting what was found by configure.
* configure.ac: Change report of configuration variables at the end of file
to use GNUPLOT_BINARY which may be "" if gnuplot was not found, even while
GNUPLOT itself has a value.

* __init_gnuplot__.cc (have_gnuplot_binary): For good measure, check that
value returned from calling gnuplot_binary is not empty before proceeding.

* gnuplot_binary.in.m: Wrap BIST tests in a try/catch block which restores
the original value of gnuplot_binary if a test fails.