movfun.m, movslice.m: Add additional input validation and BIST tests for same.
* movfun.m: Rename "basefun" to "basefcn" in documentation.
Validate number of inputs is at least 3.  Use strcmpi, rather than ismember,
in validation for performance.  Use anonymous variable @(d), not @(x) when
writing anonymous function that needs to check the existing variable 'x' so
that the value is not shadowed.  Use isindex() to verify indices for
performance.  Fix search for first non-singleton dimension so it returns a
value even for scalar or empty matrix.  Validate that WLEN is a 1- or 2-element
array of integers >= 0.  Use parfor, instead of for, for loop over columns.
Eventually this might do something meaningful in Octave.  Add BIST tests for
input validation that check the expected error message.

* movslice.m: Validate that exactly two arguments are provided.  Validate that
N is a positive integer.  Validate that WLEN is a 1- or 2-element array of
integers >= 0.  Add BIST tests for input validation that check the expected
error message.