movfun.m: Implement "discard" option for "Endpoints" property for Matlab compatibility.
* movfun.m: Add new documentation for "discard" option.  Add "discard" to
variable valid_bc.  Place call to movslice ahead of validation of Endpoints
property.  For "discard" option, change the pre-window (Cpre) and post-window
(Cpost) to be empty.  Change the indices for the centers (C) to remove the
offset of Cpre which is no longer present.  Change the length of the output
to be the length of just the centers (C).
* movfun.m (movfun_oncol): Determine the length of the output as the sum of the
lengths of Cpre, C, and Cpos (some of which may now be 0).