move variable values from symbol_record objects to stack_frame objects
Apologies for the massive commit.  I see no way to untangle these
changes into a set of smaller incremental changes in a way that would
be more useful.

Previously, handling data for recursive function calls was managed by
a stack of values in the symbol_record class and an auxiliary integer
variable was used for managing the recursion depth (context_id).  Now,
values for local variables are in the stack_frame class and recursion
is handled naturally by the call_stack as a new stack frame is added
to the call_stack object for any call to a function or a script.
Values for internal function call information (nargin, nargout, etc.)
are now stored specially in the stack_frame object.  Values for global
variables are now stored in a map in the call_stack object.  Values
for persistent variables are stored in the corresponding
octave_user_function object.  Access to non-local variables inside
nested functions is managed through pointers to stack_frame objects
for the parent function scopes.  The new implementation more closely
resembles the techniques described in standard compiler literature.

These changes should make it easier to create proper closures and
finally solve bug #39257 (handles to nested functions are not yet
supported).  They may also make it easier to implement JIT compiler,
though that is probably still a long way off.

The new stack-frame.h file has some details about the new
implementation of stack frames that should help in understanding how
things work now.

Describing each change to each file and function will probably not
provide much greater understanding of the changes and would be quite
tedious to write so I am omitting them.