Implement "none" as a color for axes and text objects (bug #52842).
* gl-render.cc (opengl_renderer::draw_axes_boxes): Check "xcolor", "ycolor",
and "zcolor" properties for special value of '[]' which means "none".  Don't
draw boxes if "color" is empty.
* gl-render.cc (opengl_renderer::draw_axes_[xyz]_grid): Check "color" property
for "none" and determine possible fallback colors for "gridcolor" and
"minorgridcolor".  Don't draw grids, axes, ticks, or ticklabels if final
color value is "none".
* gl-render.cc (opengl_renderer::draw_text): Don't draw text if "color" is
"none".
* gl-render.cc (opengl_renderer::set_color): Don't update
txt_renderer.set_color if "color" is "none".

* graphics.cc (text::properties::update_font): Don't update
txt_renderer.set_color if "color" is "none".

* graphics.in.h (axes): Add 'm' to "gridcolor" property so that changing the
color automatically flips "gridcolormode" to "manual".
* graphics.in.h (axes): Change "xcolor", "ycolor", "zcolor" properties to be a
color_property object with a possible radio_value of "none".
* graphics.in.h (text): Change "color" property to be a color_property object
with a possible radio_value of "none".
* graphics.in.h (text::update_color): Change updater function to verify color
is not "none" before proceeding.