don't use stack to save and restore lvalue list value in evaluator (bug #56752)
* pt-eval.h, pt-eval.cc (tree_evaluator::m_lvalue_list): New variable.
Reduce the overhead of saving and resstoring the pointer to the current
lvalue_list in the evaluator by using a simple pointer instead of a
stack.  The unwind_protect objects we use on the call stack already
provide the stack of values needed to save and restore the value of
the pointer to the current lvalue_list object.