improve file equaliy check for MS Windows (bug #55161).
* libinterp/corefcn/sysdep.cc (same_file_internal): Extend file equaliy check
for MS Windows system to include file size and file modification date time.
This implementation corresponds to the Boost library and to
std::filesystem::equivalent (C++17).

https://en.cppreference.com/w/cpp/filesystem/equivalent
https://www.boost.org/doc/libs/1_71_0/libs/filesystem/doc/reference.html#equivalent

The previous check was insufficient for certain MS Windows stream devices, e.g.
Google Drive File Stream.  The effect was that different files from the same
stream device were identified as equal and caused trouble when opening in the
Octave internal editor for example.