in GUI objects, store reference to base_qobject instead of resource_manager
We'll need access to both the resource_manager and the
shortcut_manager in many objects.  The shortcut_manager will be moved
to be a member of the base_qobject class instead of being a singleton,
so this change will make it easier to access both of these as well as
any other core GUI properties in any of the Octave-specific widgets
that make up the GUI.

Once again, may files are affecte but the changes are
straightforward: store a reference to the base_qobject object and
access the resource_manager object from that instead of storing a
reference to the resource_manager directly.