integrate Qt graphics toolkit in Qt interpreter object
Instead of dynamically loading Qt graphics toolkit, make it an
integral part of the Qt interpreter object.  This arrangement allows
us to pass references to both the Octave interpreter and the
base_qobject to the graphics subsystem.

Earlier, I made the Qt graphics system loadable so that it was more
like the fltk and gnuplot versions.  But that doesn't really make
sense, as the Qt graphics system requires a Qt application context and
specific support from Octave for callbacks and other things.  It can't
really be separated from Octave itself, or the parts of Octave that
are built on Qt, so it might as well be available and initialized any
time that we start a copy of Octave that is using Qt (i.e., not the
strictly "cli" version that doesn't link to the Qt libraries).

* __init_qt__.h, __init_qt__.cc: Delete.
* graphics/module.mk: Update.  Create convenience library for
libgraphics instead of __init_qt__.oct file.

* libgui/module.mk (%canon_reldir%_liboctgui_la_LIBADD): Add
libgraphics.la to the list.

* graphics-init.h, graphics-init.cc: New files.
(graphics_init): New function to initialize the graphics system.
* libgui/src/module.mk: Update.  Add -I options for libgui/graphics
directories to CPPFLAGS.

* interpreter-qobject.cc (interpreter_qobject::interpreter_qobject):
Call graphics_init to initialize graphics system.