eliminate gnulib subrepo (bug #57044)
After this change, bootstrap will check out the gnulib sources to the
revision set in the bootstrap.conf file.  You may also use

  ./bootstrap --gnulib-srcdir=/path/to/gnulib

to use an external copy of gnulib.  If GNULIB_REVISION is unset, then
the external repo will be updated to the version specified in
bootstrap.conf.  If GNULIB_REVISION is set in the environment when
running bootstrap, that version will be used.  If GNULIB_REVISION is
set but empty, the current version in the external repo will be used:

  GNULIB_REVISION= ./bootstrap --gnulib-srcdir=/path/to/gnulib

Thanks to Kai T. Ohlhus for the initial version of this change.

* bootstrap.conf: Set GNULIB_REVISION unless it is already set in the
environment.
* .hgsub, .hgsubstate: Delete.