inputname.m: Closer compatibility with Matlab for special values (i,j,NaN,Inf) (bug #59103)
* variables.cc (F__varval__): Add more documentation to internal function.
Use same name for variables in documentation and in code for understanding.

* inputname.m: Rewrite documentation to further explain the function.
Add FIXME notes about ways in which inputname() is still different from
Matlab's implementation.  Add tighter input validation of index input N.
Add second-stage of testing for variable names that *might* be valid,
but might also be special values.  Rewrite BIST tests so it does not depend
on internal behavior of test.m function which might change in the future.
Add new BIST tests for function and for input validation.