also wait on main_thread after interpreter shuts down (bug #56952)
I'm not sure whether this change will fix the crash at exit bug, but
it is documented that a call to QThread::quit should be followed by a
call to QThread::wait.

* octave-qobject.cc
(base_qobject::handle_interpreter_shutdown_finished): As recommended
in the Qt docs, follow the call to QThread::quit on the m_main_thread
object with a call to QThread::quit.