don't clear dock widget pointers in main window dtor (bug #60646, bug #60750)
If a dock widget is docked to the main window, it will have the main
window as its parent and will be deleted when the main_window object
is deleted.  Then, because we use QPointer for those objects, it will
be set to nullptr and the base_qobject will recreate it if needed.  If
a dock widget is not docked, then its parent will not be main_window
and it will not be deleted when the main_window destructor is
executed and will remain available if Octave is returning to command
line mode.

* main-window.cc (main_window::~main_window):
Don't clear dock widget pointers.