Change startup so that '--interactive' does not always force '--no-line-editing' (bug #60696).
* interpreter.cc (interpreter::interpreter): Create temporary variable "forced_interactive"
that holds state from the const-declared variable "options".  Check if forced interactive
option was given in conjunction with a session that is already interactive ("m_interactive").
If found, then reset the forced nature of session to false both in the application ("app_context")
and locally in the temporary variable "forced_interactive".