idivide.m: Avoid floating-point representation issues (bug #61319).
* scripts/general/idivide.m:
We start from `z = x ./ y` which corresponds to the classical rounding, and
which we are sure is an integer value if `x` or `y` is an integer value. We
then add or subtract `1` if necessary according to the rounding rule and the
sign of `y`, keeping the starting integer type all along.
Simplify input validation. Add BISTs.