textscan: Reposition scan pointer after reading a single `.` (bug #60711).
* oct-stream.cc (textscan::read_double):
When a single `.` is encountered, it is indeed considered an invalid character.
But it must also be put back because, unlike any other invalid character, it
has caused an additional `get()` call above.
This then allows the various tests for the presence of a delimiter on the next
character to fail (e.g., in `textscan::read_format_once()` and
`textscan_format_list::read_first_row()`), and thus finally allows `textscan()`
builtin to exit correctly in error.

* file-io.cc (Ftextscan): Add BIST.

* doc/interpreter/contributors.in: Add new contributor to list in manual.