eliminate range<T>::make_constant function
Overloading ranges as a way to create constant arrays adds complexity
to the range class for limited benefit.  It is currently only used for
double-valued ranges and only provides constant row vectors.  If this
optimization is determined to be useful, it would probably be best to
provide a data type that allows N-d constant arrays.  The legacy range
class was revived in a previous changeset so that constant ranges
(increment == 0) can be loaded from existing data files and converted
to ordinary arrays automatically.

* Range.h (range<T>::make_constant): Delete.  Remove all uses.