doc: Update refcard.tex content from Octave 3 to Octave 8
refcard.tex: Remove Readline material, expand Octave examples.
The intended audience of this reference card is someone who
already knows programming in a different language like Python
and is referring to this refcard.ps / refcard.pdf for Octave
syntax. It is not a comprehensive reference, but gives enough
function names to be useful with help / doc.
Intended length 2 pages.