Mark exceptions with default visibility also where they are used (bug #61704).
For exceptions that are thrown across library borders to work correctly, the
definition of those exceptions needs to be attributed with default visibility
both in the library where they are defined and in the code where they are
caught.

* oct-conf-post-public.in.h: Add new preprocessor macros for visibility
attributes of exception classes.
* liboctave/util/quit.h (execution_exception),
liboctave/util/lo-array-errwarn.h (index_exception): Use new preprocessor macro
to flag exception classes that are thrown across library borders.