Overhaul scanf functionality for correctness and Matlab compatibility (bug #63383, bug #62723) Return correct third output (error message) and fourth output (position) in many more cases.  Functionality is still not correct in all situations and new corner cases were discovered while coding.  Add BIST tests for correct behavior, some of which fail still.
* oct-stream.cc (BEGIN_C_CONVERSION, BEGIN_S_CONVERSION): After reading
characters, clear the failbit if the last character that was attempted to be
read was EOF.  This is not a true failure.
* oct-stream.cc (BEGIN_CHAR_CLASS_CONVERSION): Move initialization of "width"
variable to be with its definition as int.  When reading characters and the
read character is not in the character class, call putback(c) to place
the character back into the input stream (cause of misplaced position output).
After reading characters, clear the failbit if the last character that was
attempted to be read was EOF.  This is not a true failure.
* oct-stream.cc (FINISH_CHARACTER_CONVERSION): Check for "width > 0", i.e.,
there were characters read, before updating outputs like conversion_count and
position.
* oct-stream.cc (base_stream::do_scanf): Remove test for "! eof" and just
check for failbit.  If failbit is set then issue a warning about pattern
failing to match.

* lo-utils.cc (read_fp_value): If reading has failed (failbit set) then clear
eofbit which may have been set during the process.  The calling routine will
use seek() to restore file position to point before attempted read so EOF
should no longer be set on stream.

* io.tst: Add many more BIST cases.