print: Add option to merge only subsequent triangles with SVG toolchain (bug #63646).
Trying to merge all triangles that are sharing an edge might take a long time.
But that might be necessary to avoid hairlines in figures containing patch or
surface graphics objects. Allow selecting if no, all, or only subsequent
triangles sharing an edge should be merged into a polygon.

* src/octave-svgconvert.cc (main): Distinguish between merging no (0), only
consecutive (1), or all (2) triangles into polygons.
(octave_polygon::reconstruct): Skip numerically expensive part of trying to
merge all polygons that are sharing an edge unless it was selected.

* scripts/plot/util/print.m: Document new options "-polymerge", "-nopolymerge",
and "-polymerge-all" for polygon merging with the SVG toolchain.
(svgconvert): Call "octave-svgconvert" with the selected polygon merge mode.

* scripts/plot/private/__print_parge_opts__.m: Select default value and parse
input for new options.