movfun.m: Return correct class of output (bug #63802).
* movfun.m: Check whether there are any elements to process based on requested
dimension.  If there are none, short-circuit execution and return an empty
matrix with the same size as the input.  Rename variable "tmp" to "fout" for
clarity.  Declare new variable "yclass" which is the class ouf the output
determined by executing "f(x)".  Initialize output "y" to be an array of
zeros of class "yclass".  Add BIST tests for bug #63802.
* movfun.m (movfun_oncol): Add new input parameter "yclass" to function.
Initialize output "y" to be an array of zeros of class "yclass".