correctly load all-zero sparse matrices from text files (bug #64696)
Since changeset c0d8ce61c1c9, sparse arrays have reserved at least one
element of storage so nzmax is always greater than 0.  That causes
trouble for the read_sparse_matrix function which only reads a set of
triplets (ROW, COL, VAL) from the input stream and assumes that the
sparse array that it receives and fills with values has been
initialized with exactly the correct amount of storage to hold all the
values in the file (it assumes that nnz is the same as nzmax).  But
that's not the case for sparse arrays created with nnz equal to 0 and
read_sparse_matrix attempts to read an element when it should not.  We
can't avoid the problem by calling Sparse<T>::nnz in
read_sparse_matrix because it returns zero for a newly allocated
sparse array and won't be correct until the values are stored in the
array.  Instead, we have to skip reading array elements in the
load_ascii function if NZ is zero.

* ov-base-sparse.cc (octave_base_sparse<T>::load_ascii): Skip input of
matrix elements if NNZ is zero.  New test.