symrcm.cc: Speed up degree calculation (Bug #64718)
* symrcm.cc: Rework calculation of vertex degrees for a sparse matrix
to make it faster. Previous version was doing double-work with transpose.
New version replaces it with twice the degree, which does not affect the
algorithm but is simpler and faster. For 5000 x 5000 input, runtime is now
some 2.7 seconds instead of 258 seconds, and scales quadratically as expected
instead of cubically as before. Also move early return paths earlier
in the DEFUN, and return 1:N instead of 0:(N-1) for all-zeros input.