Make random numbers in BISTs reproducible
Comparing fntests.log across different runs was difficult due to
large random arrays being printed for some known failures. This change
sets the random state for such instances that were being printed.
It also sets the random state as a non-shared variable so that it is
not printed, which would again cause difficulty comparing test runs.

* libinterp/corefcn/conv2.cc : set state
* lininterp/corefcn/gsvd.cc : set state