New opcode for i:s and e
Specialized opcode that deals with 'i','j','I','J' and 'e' when used as
the imaginary unit or natural log base, to push faster. As with PUSH_PI.

Being assigned to is used as a heuristic for if the id:s are used
as variables or as mathematical constants. A fallback covers the cases
where the guess is wrong or the user has overridden the builtin functions.

* pt-bytecode-vm.cc: New opcodes
* pt-bytecode-vm.h: New fields for checking for overrides
* pt-bytecode-walk.cc: Emit new opcodes
* pt-bytecode-walk.h: New field m_set_assigned_ids
* pt-bytecode.h: New opcodes PUSH_I, PUSH_E